退款功能

+++ trade/src/main/java/com/meirengu/trade/model/Order.java
@@ -1,35 +1,36 @@
 package com.meirengu.trade.model;
 
 import java.util.Date;
+import java.math.BigDecimal;
  /*
   * Order 实体类
   * Tue Mar 14 17:15:51 CST 2017 建新
   */
 public class Order{
 	/** 众筹订单索引id */
-	private int orderId;
+	private Integer orderId;
 	/** 众筹订单编号，内部使用编号 */
 	private String orderSn;
 	/** 众筹项目id */
-	private int itemId;
+	private Integer itemId;
 	/** 众筹项目名称 */
 	private String itemName;
 	/** 众筹档位id */
-	private int itemLevelId;
+	private Integer itemLevelId;
 	/** 档位名称 */
 	private String itemLevelName;
 	/** 档位价格 */
-	private double itemLevelAmount;
+	private BigDecimal itemLevelAmount;
 	/** 众筹份数 */
-	private int itemNum;
+	private Integer itemNum;
 	/** 订单总价格 */
-	private double orderAmount;
+	private BigDecimal orderAmount;
 	/** 本金（本金=订单总价格-抵扣券金额） */
-	private double costAmount;
+	private BigDecimal costAmount;
 	/** 抵扣券金额 */
-	private double rebateAmount;
+	private BigDecimal rebateAmount;
 	/** 买家id */
-	private int userId;
+	private Integer userId;
 	/** 买家姓名 */
 	private String userName;
 	/** 买家手机号 */
@@ -37,11 +38,11 @@ public class Order{
 	/** 买家电子邮箱 */
 	private String userEmail;
 	/** 买家收货地址 */
-	private int userAddressId;
+	private Integer userAddressId;
 	/** 订单生成时间 */
 	private Date createTime;
 	/** 订单类型 1.普通 2.其他 */
-	private int orderType;
+	private Integer orderType;
 	/** 订单编号，外部支付时使用，有些外部支付系统要求特定的订单编号 */
 	private String outSn;
 	/** 订单完成时间 */
@@ -51,9 +52,9 @@ public class Order{
 	/** 订单留言 */
 	private String orderMessage;
 	/** 订单状态：1预约待审核；2预约审核通过（前端显示未付款状态）3预约审核未通过（前端显示已关闭状态） 4未付款；5已失效（前端显示已关闭状态） 6已付款 7未消费失效 8已消费  9申请退款  10确认退款  11拒绝退款 12退款成功 13支付失败 */
-	private int orderState;
+	private Integer orderState;
 	/** 逻辑删除状态 0为删除，1为未删除 */
-	private int flag;
+	private Integer flag;
 	/** 1:pc，2:h5，3:ios，4:android，5：三方 */
 	private String orderFrom;
 	/** 修改时间 */
@@ -63,11 +64,11 @@ public class Order{
 	/** 用户微信号 */
 	private String userWeixin;
 
-	public void setOrderId(int orderId){
+	public void setOrderId(Integer orderId){
 		this.orderId = orderId;
 	}
 
-	public int getOrderId(){
+	public Integer getOrderId(){
 		return orderId;
 	}
 
@@ -79,11 +80,11 @@ public class Order{
 		return orderSn;
 	}
 
-	public void setItemId(int itemId){
+	public void setItemId(Integer itemId){
 		this.itemId = itemId;
 	}
 
-	public int getItemId(){
+	public Integer getItemId(){
 		return itemId;
 	}
 
@@ -95,11 +96,11 @@ public class Order{
 		return itemName;
 	}
 
-	public void setItemLevelId(int itemLevelId){
+	public void setItemLevelId(Integer itemLevelId){
 		this.itemLevelId = itemLevelId;
 	}
 
-	public int getItemLevelId(){
+	public Integer getItemLevelId(){
 		return itemLevelId;
 	}
 
@@ -111,51 +112,51 @@ public class Order{
 		return itemLevelName;
 	}
 
-	public void setItemLevelAmount(double itemLevelAmount){
+	public void setItemLevelAmount(BigDecimal itemLevelAmount){
 		this.itemLevelAmount = itemLevelAmount;
 	}
 
-	public double getItemLevelAmount(){
+	public BigDecimal getItemLevelAmount(){
 		return itemLevelAmount;
 	}
 
-	public void setItemNum(int itemNum){
+	public void setItemNum(Integer itemNum){
 		this.itemNum = itemNum;
 	}
 
-	public int getItemNum(){
+	public Integer getItemNum(){
 		return itemNum;
 	}
 
-	public void setOrderAmount(double orderAmount){
+	public void setOrderAmount(BigDecimal orderAmount){
 		this.orderAmount = orderAmount;
 	}
 
-	public double getOrderAmount(){
+	public BigDecimal getOrderAmount(){
 		return orderAmount;
 	}
 
-	public void setCostAmount(double costAmount){
+	public void setCostAmount(BigDecimal costAmount){
 		this.costAmount = costAmount;
 	}
 
-	public double getCostAmount(){
+	public BigDecimal getCostAmount(){
 		return costAmount;
 	}
 
-	public void setRebateAmount(double rebateAmount){
+	public void setRebateAmount(BigDecimal rebateAmount){
 		this.rebateAmount = rebateAmount;
 	}
 
-	public double getRebateAmount(){
+	public BigDecimal getRebateAmount(){
 		return rebateAmount;
 	}
 
-	public void setUserId(int userId){
+	public void setUserId(Integer userId){
 		this.userId = userId;
 	}
 
-	public int getUserId(){
+	public Integer getUserId(){
 		return userId;
 	}
 
@@ -183,11 +184,11 @@ public class Order{
 		return userEmail;
 	}
 
-	public void setUserAddressId(int userAddressId){
+	public void setUserAddressId(Integer userAddressId){
 		this.userAddressId = userAddressId;
 	}
 
-	public int getUserAddressId(){
+	public Integer getUserAddressId(){
 		return userAddressId;
 	}
 
@@ -199,11 +200,11 @@ public class Order{
 		return createTime;
 	}
 
-	public void setOrderType(int orderType){
+	public void setOrderType(Integer orderType){
 		this.orderType = orderType;
 	}
 
-	public int getOrderType(){
+	public Integer getOrderType(){
 		return orderType;
 	}
 
@@ -239,19 +240,19 @@ public class Order{
 		return orderMessage;
 	}
 
-	public void setOrderState(int orderState){
+	public void setOrderState(Integer orderState){
 		this.orderState = orderState;
 	}
 
-	public int getOrderState(){
+	public Integer getOrderState(){
 		return orderState;
 	}
 
-	public void setFlag(int flag){
+	public void setFlag(Integer flag){
 		this.flag = flag;
 	}
 
-	public int getFlag(){
+	public Integer getFlag(){
 		return flag;
 	}
